Refactor core utility functions and evaluation logic
