+++ src/js/core/coreUtils.js
